refactor(api): extracts prompt-chat usecase from llm-api
